+++ index.html
@@ -3,8 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
-    <title>Document</title>
+    <link rel="stylesheet" href="./css/style.css">
+    <title>Event Page</title>
 
 </head>
 <body>
@@ -19,7 +19,7 @@
     <div class="event-grid"></div>
   </div>
 
-  <script src="main.js"></script>
+  <script type="module" src="./js/main.js"></script>
 </body>
 </html>
